Consistent use of int in mlir_runner_utils.cpp
This should fix the OSS build by only using int in template types.

PiperOrigin-RevId: 274843584

+++ test/mlir-cpu-runner/mlir_runner_utils.cpp
@@ -24,7 +24,7 @@
 #include <assert.h>
 #include <iostream>
 
-template <size_t... Dims> struct StaticSizeMult {
+template <int... Dims> struct StaticSizeMult {
   static constexpr int value = 1;
 };
